mejore la arquitectura de Aprender.tsx
